Move `CommittableOffset` to `ConsumerHandler`
